requirements.txt
fastapi[all]
uvicorn
sqlalchemy
passlib
bcrypt==3.2.0
python-multipart
pyjwt
python-dotenv
aiofiles
requests
pytest
jinja2
